python/requirements.txt
flet==0.12.0
tensorflow
tensorflow-lite
pillow
numpy